Discrete_SSL/requirements.txt
# Core dependencies
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.21.0
scikit-learn>=1.0.0

# SpeechBrain for SSL quantization
speechbrain>=1.0.0

# Transformers and HuggingFace
transformers>=4.30.0
huggingface-hub>=0.16.0

# Data processing
pandas>=1.3.0
PyYAML>=5.4.0
joblib>=1.0.0

# Training utilities
tqdm>=4.62.0

# Experiment tracking
wandb>=0.12.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Audio processing (optional)
librosa>=0.9.0
soundfile>=0.10.0